BLI: Add normalized integer classes
This allow easier support for packed special vertex
and texture formats.

I kept the same non-standard compliant approach
as the `PackedNormal` implementation we already
had.

The `convert_normal` was not even consistent,
it was clamping for the low bitdepth but not
the high bit depth.

In order to be consistent and avoid performance
regressions, I removed the clamping. We can
move that responsibility to the caller.

Rel #130640

+++ source/blender/gpu/tests/vertex_buffer_test.cc
@@ -128,7 +128,7 @@ GPU_TEST(vertex_buffer_fetch_mode__GPU_COMP_U16__GPU_FETCH_INT_TO_FLOAT_UNIT);
 static void test_vertex_buffer_fetch_mode__GPU_COMP_I10__GPU_FETCH_INT_TO_FLOAT_UNIT()
 {
   vertex_buffer_fetch_mode<VertAttrType::SNORM_10_10_10_2, PackedNormal>(
-      PackedNormal(321, -511, 511, 0));
+      PackedNormal(int4(321, -511, 511, 0)));
 }
 GPU_TEST(vertex_buffer_fetch_mode__GPU_COMP_I10__GPU_FETCH_INT_TO_FLOAT_UNIT);
